% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/MatrixLayers-class.R
\name{defaultLayer}
\alias{defaultLayer}
\alias{defaultLayer<-}
\alias{defaultLayer.MatrixLayers}
\alias{defaultLayer<-.MatrixLayers}
\title{Get or set the default layer}
\usage{
defaultLayer(object, ...)

defaultLayer(object, ...) <- value

\method{defaultLayer}{MatrixLayers}(object, ...)

\method{defaultLayer}{MatrixLayers}(object, ...) <- value
}
\arguments{
\item{object}{A \code{\link{MatrixLayers}} object.}

\item{...}{Arguments passed to other metheds.}

\item{value}{A character or an integer specifying which layer should be the
new default. For a character, layers in \code{x} must contain names.}
}
\value{
\itemize{
\item \code{defaultLayer}: A single integer specifying which layer is the default.
}

\itemize{
\item \verb{defaultLayer<-}: \code{object} with updated default layer.
}
}
\description{
Methods to get or set the index of default layer for a
\code{\link{MatrixLayers}}.
}
