% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/joinLayers.R
\name{joinLayers}
\alias{joinLayers}
\alias{joinLayers.MatrixLayers}
\title{Join matrix layers}
\usage{
joinLayers(object, ...)

\method{joinLayers}{MatrixLayers}(object, by = c("column", "row"), ...)
}
\arguments{
\item{object}{A \code{\link{MatrixLayers}} object.}

\item{...}{\code{.dot_param}}

\item{by}{Either 'column' or 'row'. Which axis should the layers be joined
along with.}
}
\value{
A joined matrix-like object, whose row names and column names are ordered
according to the dimension maps.
}
\description{
Join matrix layers together, by either rows or columns.
}
\details{
The underlying implemention is to call \code{\link{concatByRows}} or
\code{\link{concatByCols}} on the matrix layers according to \code{by}. Therefore,
a \code{\link{MatrixLayers}} containing matrix-like objects that has
implemented these two methods should work when calling \code{joinLayers}.

If \code{by} is set to 'row' for a \code{\link{MatrixLayers}}, no intersection is
allowed in the row map. If \code{by} is 'column', no intersection is allowed in
the column map.
}
