% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLayers-class.R
\name{length.MatrixLayers}
\alias{length.MatrixLayers}
\title{Number of layers}
\usage{
\method{length}{MatrixLayers}(x)
}
\arguments{
\item{x}{A \code{\link{MatrixLayers}} object.}
}
\value{
An integer indicating the number of layers.
}
\description{
Methods to get the number of layers in a \code{\link{MatrixLayers}}.
}
