% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{Matrix.Layered-package}
\alias{Matrix.Layered}
\alias{Matrix.Layered-package}
\title{Matrix.Layered: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
