% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/MatrixLayers-class.R
\name{MatrixLayers-accessors}
\alias{MatrixLayers-accessors}
\alias{getLayer}
\alias{setLayer}
\alias{getLayers}
\alias{setLayers}
\alias{getLayer.MatrixLayers}
\alias{setLayer.MatrixLayers}
\title{Get or set layer data}
\usage{
getLayer(object, i, ...)

setLayer(object, i, new.layer, ...)

getLayers(object, layers, ...)

setLayers(object, layers, new.layers, ...)

\method{getLayer}{MatrixLayers}(object, i, ...)

\method{setLayer}{MatrixLayers}(object, i, new.layer, ...)
}
\arguments{
\item{object}{A \code{\link{MatrixLayers}} object.}

\item{i}{A single character or integer to specify the layer to get or set.}

\item{...}{Arguments passed to other metheds.}

\item{new.layer}{A matrix-like object to add as a new layer \code{i}.}

\item{layers}{A vector of characters or integers to specify which layers to
get or set.}

\item{new.layers}{A list of matrix-like objects to set. Must has the same
length as \code{layers}}
}
\value{
\itemize{
\item \code{getLayer}: A matrix-like object for layer \code{i}.
}

\itemize{
\item \code{setLayer}: An updated \code{object} with \code{value} saved as layer \code{i}
}

\itemize{
\item \code{getLayers}: A list of fetched layers.
}

\itemize{
\item \code{setLayers}: An updated \code{object} with \code{layers} are set or replaced with
\code{new.layers}.
}
}
\description{
Methods to get or set layer data for a \code{\link{MatrixLayers}}.
}
