% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLayers-class.R
\name{MatrixLayers-dim}
\alias{MatrixLayers-dim}
\alias{dim.MatrixLayers}
\alias{dimnames.MatrixLayers}
\alias{dimnames<-.MatrixLayers}
\title{Dimensions of MatrixLayers}
\usage{
\method{dim}{MatrixLayers}(x)

\method{dimnames}{MatrixLayers}(x)

\method{dimnames}{MatrixLayers}(x) <- value
}
\arguments{
\item{x}{A \code{\link{MatrixLayers}} object.}

\item{value}{A two-element list, where the first character vector specifies
new row names and the second specifies new column names.}
}
\value{
\itemize{
\item \code{dim}: A two-element integer vector indicating the numbers of rows
and columns for \code{x}.
}

\itemize{
\item \code{dimnames}: A two-element character list indicating the row and
column names for \code{x}.
}

\itemize{
\item \verb{dimnames<-}: An updated \code{x} with new row and column names.
}
}
\description{
Methods to get or set the dimension names for a \code{\link{MatrixLayers}}.
}
