% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLayers-class.R
\name{MatrixLayers-dollar}
\alias{MatrixLayers-dollar}
\alias{.DollarNames.MatrixLayers}
\alias{$.MatrixLayers}
\alias{$<-.MatrixLayers}
\title{Get or set layer data}
\usage{
\method{.DollarNames}{MatrixLayers}(x, pattern = "")

\method{$}{MatrixLayers}(x, i)

\method{$}{MatrixLayers}(x, i) <- value
}
\arguments{
\item{x}{A \code{\link{MatrixLayers}} object}

\item{pattern}{A regular expression. Only matching names are returned.}

\item{i}{Name of layer data to get or set}

\item{value}{A matrix-like object to add as a new layer \code{i}.}
}
\value{
\itemize{
\item \code{.DollarNames}: The layer name matches for \code{pattern}.
}

\itemize{
\item \code{$}: A matrix-like object for layer \code{i}.
}

\itemize{
\item \verb{$<-}: An updated \code{x} with \code{value} saved as layer \code{i}
}
}
\description{
Getting or setting layer data for a \code{\link{MatrixLayers}} using `$.
}
\seealso{
\code{\link[utils:rcompgen]{utils::.DollarNames()}}
}
\keyword{internal}
