% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLayers-class.R
\name{MatrixLayers-names}
\alias{MatrixLayers-names}
\alias{names.MatrixLayers}
\alias{names<-.MatrixLayers}
\title{Get or set layer names}
\usage{
\method{names}{MatrixLayers}(x)

\method{names}{MatrixLayers}(x) <- value
}
\arguments{
\item{x}{A \code{\link{MatrixLayers}} object.}

\item{value}{The new layer namse to set.}
}
\value{
\itemize{
\item \code{names}: A character vector specifying layer names.
}

\itemize{
\item \verb{names<-}: An updated \code{x} with new layer names.
}
}
\description{
Methods to get or set layer names for a \code{\link{MatrixLayers}}.
}
