% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/logical_map-nsparseMatrix.R
\name{sparseLogMap}
\alias{sparseLogMap}
\alias{sparseLogMap.character}
\title{Create a sparse logical map}
\usage{
sparseLogMap(x, ...)

\method{sparseLogMap}{character}(x, ncol = 0, names = NULL, ...)
}
\arguments{
\item{x}{A character vector}

\item{...}{Arguments passed to other metheds.}

\item{ncol}{Number of column (observation) for the output matrix.}

\item{names}{Names of observations. Must have \code{ncol} elements.}
}
\value{
An \code{\link[Matrix:nsparseMatrix-class]{nsparseMatrix}} with row number equal to length of \code{x},
representing whether each value (row) in \code{x} exists in each observation
(colum).
}
\description{
Create a container for storing mappings of values using sparse logical
matrices. Each row specifies whether a value (row name) is present in which
observations (columns).
}
\examples{
# Create a sparse logical map
map <- sparseLogMap(LETTERS)

# Create a logical map for known observations
obs <- letters[1:4]
map <- sparseLogMap(LETTERS, ncol = length(obs), names = obs)
map

}
